+++ Dockerfile
@@ -2,11 +2,12 @@ FROM golang:alpine
 
 RUN apk update && apk add git
 
-RUN mkdir /app
-
+# ワーキングディレクトリの設定
 WORKDIR /app
 
-COPY ./app .
+# ホストのファイルをコンテナの/appにコピーする
+COPY * ./
 
+# airを使ってホットリロード
 RUN go get -u github.com/cosmtrek/air
 CMD ["air", "-c", ".air.toml"]
