+++ models/base.go
@@ -1,0 +1,33 @@
+package models
+
+import (
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/jinzhu/gorm"
+	_ "github.com/jinzhu/gorm/dialects/mysql"
+	"github.com/joho/godotenv"
+)
+
+var DB *gorm.DB
+
+func init() {
+	err := godotenv.Load(".env")
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	mysqlUser := os.Getenv("MYSQL_USER")
+	mysqlPassword := os.Getenv("MYSQL_PASSWORD")
+	mysqlHost := os.Getenv("MYSQL_HOST")
+	mysqlPort := os.Getenv("MYSQL_PORT")
+	mysqlDatabase := os.Getenv("MYSQL_DATABASE")
+	dsn := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s?parseTime=true", mysqlUser, mysqlPassword, mysqlHost, mysqlPort, mysqlDatabase)
+	DB, err = gorm.Open("mysql", dsn)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	DB.AutoMigrate(&Post{}) // postsテーブルの作成
+}
