+++ models/post.go
@@ -1,0 +1,10 @@
+package models
+
+import "time"
+
+type Post struct {
+	ID        int       `json:"id"`
+	Content   string    `json:"content"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
